+++ packages/privacy-manager/src/components/PrivacyManagerForm.tsx
@@ -1,0 +1,27 @@
+import React, { useCallback } from 'react'
+import { PrivacyManagerDecisions } from '../config'
+import { useConfig } from '../context'
+import { useEnabledIntegrations } from '../integrations'
+import { DecisionsForm, DecisionsFormState } from './DecisionsForm'
+
+export const PrivacyManagerForm: React.FC<{
+  onSubmit?: (decisions: PrivacyManagerDecisions) => void
+}> = () => {
+  const [enabledIntegrations, setEnabledIntegrations] = useEnabledIntegrations()
+  const { integrations } = useConfig()
+
+  const handleFormSubmit = useCallback(
+    (values: DecisionsFormState) => {
+      setEnabledIntegrations(() => values.enabled)
+    },
+    [setEnabledIntegrations]
+  )
+
+  return (
+    <DecisionsForm
+      integrations={integrations}
+      intitialValues={{ enabled: enabledIntegrations }}
+      onSubmit={handleFormSubmit}
+    />
+  )
+}